+++ netbox_nsm/api/serializers_/type_config.py
@@ -4,6 +4,7 @@ from rest_framework.serializers import HyperlinkedIdentityField
 from netbox.api.fields import ContentTypeField
 from netbox.api.serializers import NetBoxModelSerializer
 
+from netbox_nsm.changelog_utils import apply_type_config_changelog_message
 from netbox_nsm.models import TypeConfig
 
 __all__ = ("TypeConfigSerializer",)
@@ -52,3 +53,17 @@ class TypeConfigSerializer(NetBoxModelSerializer):
             "panel_slugs",
             "panel_linkable_types",
         )
+
+    def update(self, instance, validated_data):
+        prechange = getattr(instance, "_prechange_snapshot", None)
+        if prechange and "changelog_message" not in validated_data:
+            for attr, value in validated_data.items():
+                if attr != "changelog_message":
+                    setattr(instance, attr, value)
+            apply_type_config_changelog_message(instance, prechange=prechange)
+            if instance._changelog_message:
+                validated_data = {
+                    **validated_data,
+                    "changelog_message": instance._changelog_message,
+                }
+        return super().update(instance, validated_data)